raise NoLyricsError if data indicates missing lyrics
The song info endpoint (e.g.
/api.php?artist=Al+Green&action=lyrics&song=Let%27+Stay+Together&fmt=realjson)
response includes a field `lyrics` containing the first few lines of the
song's lyrics. If the song's lyrics are missing, the field will contain
the string "Not found". In that case, don't bother trying to fetch any
more information about the song (n.b. the URL in the response data can
be the edit page on the wiki).